+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Changed hardcoded `px`, or `rem` value in components to spacing scalable alternatives
 - Split uncompiled CSS into many different files for easier readability
+- Changed `range`utility to be configurable and end-exclusive
 
 ### Fixed